Refactor comment.dart and post_detail_screen.dart
This commit refactors the comment.dart file to update the vote and parent properties to accept nullable integer values. It also refactors the post_detail_screen.dart file to import necessary dependencies and add a BlocBuilder widget to display comments based on the state of the DetailPostBloc. These changes improve the code structure and enhance the functionality of the app by allowing users to view and interact with comments on posts.

+++ lib/data/models/comment.dart
@@ -84,13 +84,13 @@ class Comment {
       upvotes: map['upvotes'].toInt() as int,
       downvotes: map['downvotes'].toInt() as int,
       score: map['score'].toInt() as int,
-      vote: map['vote'].toInt() as int?,
+      vote: map['vote'] as int?,
       content: map['content'] as String,
       created_at: map['created_at'] as String,
       updated_at: map['updated_at'] as String,
       depth: map['depth'].toInt() as int,
       post: map['post'].toInt() as int,
-      parent: map['parent'].toInt() as int?,
+      parent: map['parent'] as int?,
     );
   }
